+++ src/index.js
@@ -1,4 +1,5 @@
+import ObjectValueWrapper from './js/components/ObjectValueWrapper';
 import DualListBox from './js/DualListBox';
 
-export { DualListBox };
+export { ObjectValueWrapper, DualListBox };
 export default DualListBox;